Crate query_engine_tests::query_core

Modules

constants

executor

What the executor module DOES:

protocol

query_document

Intermediate representation of the input document that is used by the query engine to build query ASTs and validate the incoming data.

query_graph_builder

Query graph builder module.

response_ir

Prisma Response IR (Intermediate Representation).

schema

schema_builder

Query schema builder. Root for query schema building.

telemetry

Structs

BatchDocumentTransaction

CompactedDocument

ExtendedTransactionUserFacingError

FieldConversionError

In

Selection

TransactionOptions

TxId

How Interactive Transactions work The Interactive Transactions (iTx) follow an actor model design. Where each iTx is created in its own process. When a prisma client requests to start a new transaction, the Transaction Actor Manager spawns a new ITXServer. The ITXServer runs in its own process and waits for messages to arrive via its receive channel to process. The Transaction Actor Manager will also create an ITXClient and add it to hashmap managed by an RwLock. The ITXClient is the only way to communicate with the ITXServer. Once the prisma client receives the iTx Id it can perform database operations using that iTx id. When an operation request is received by the TransactionActorManager, it looks for the client in the hashmap and passes the operation to the client. The ITXClient sends a message to the ITXServer and waits for a response. The ITXServer will then perform the operation and return the result. The ITXServer will perform one operation at a time. All other operations will sit in the message queue waiting to be processed.

Enums

ArgumentValue

Represents the input values in a Document. This abstraction is mostly there to hold special kind of values such as FieldRef which have to be disambiguated at query-validation time.

BatchDocument

CoreError

Operation

QueryDocument

SelectionSet

TransactionError

Traits

QueryExecutor

Type Definitions

ArgumentValueObject

Result

Result type tying all sub-result type hierarchies of the core together.

SelectionArgument