Struct query_engine_tests::query_core::query_graph::QueryGraph

pub struct QueryGraph { /* private fields */ }

A graph representing an abstract view of queries and their execution dependencies.

Graph invariants (TODO put checks into the code?):

Directed, acyclic.

Node IDs are unique and stable.

The graph may have multiple result nodes, and multiple paths in the graph may point to result nodes, but only one result is serialized. Note: The exact rules determining the final result are subject of the graph translation.

Currently, Nodes are allowed to have multiple parents, but the following invariant applies: They may only refer to their parent and / or one of its ancestors. Note: This rule guarantees that the dependent ancestor node result is always in scope for fulfillment of dependencies.

Following the above, sibling dependencies are disallowed as well.

Edges are ordered. Node: Their evaluation is performed from low to high ordering, unless other rules require reshuffling the edges during translation.

Implementations

impl QueryGraph

pub fn new() -> QueryGraph

pub fn root<F>(f: F) -> Result<QueryGraph, QueryGraphBuilderError> where
F: FnOnce(&mut QueryGraph) -> Result<(), QueryGraphBuilderError>,

pub fn finalize(&mut self) -> Result<(), QueryGraphError>

pub fn find_result_node(&self, starting_node: &NodeRef) -> Option<NodeRef>

Returns a NodeRef to the result node that occurs in the subtree, if it exists. Returns None if no such node is found.

pub fn result_nodes(&self) -> Vec<NodeRef, Global>

pub fn add_result_node(&mut self, node: &NodeRef)

Adds a result node to the graph.

pub fn mark_visited(&mut self, node: &NodeRef)

pub fn is_result_node(&self, node: &NodeRef) -> bool

Checks if the given node is marked as one of the result nodes in the graph.

pub fn subgraph_contains_result(&self, node: &NodeRef) -> bool

Checks if the subgraph starting at the given node contains the node designated as the overall result.

pub fn root_nodes(&self) -> Vec<NodeRef, Global>

Returns all root nodes of the graph. A root node is defined by having no incoming edges.

pub fn create_node<T>(&mut self, t: T) -> NodeRef where
T: Into<Node>,

Creates a node with content t and adds it to the graph. Returns a NodeRef to the newly added node.

pub fn create_edge(
&mut self,
from: &NodeRef,
to: &NodeRef,
content: QueryGraphDependency
) -> Result<EdgeRef, QueryGraphError>

Creates an edge with given content, originating from node from and pointing to node to. Checks are run after edge creation to ensure validity of the query graph. Returns an EdgeRef to the newly added edge. Todo currently panics, change interface to result type.

pub fn flag_transactional(&mut self)

Mark the query graph to need a transaction.

pub fn needs_transaction(&self) -> bool

If true, the graph should be executed inside of a transaction.

pub fn node_content(&self, node: &NodeRef) -> Option<&Node>

Returns a reference to the content of node, if the content is still present.

pub fn edge_content(&self, edge: &EdgeRef) -> Option<&QueryGraphDependency>

Returns a reference to the content of edge, if the content is still present.

pub fn edge_source(&self, edge: &EdgeRef) -> NodeRef

Returns the node from where edge originates (e.g. source).

pub fn edge_target(&self, edge: &EdgeRef) -> NodeRef

Returns the node to which edge points (e.g. target).

pub fn outgoing_edges(&self, node: &NodeRef) -> Vec<EdgeRef, Global>

Returns all edges originating from= node (e.g. outgoing edges).

pub fn incoming_edges(&self, node: &NodeRef) -> Vec<EdgeRef, Global>

Returns all edges pointing to node (e.g. incoming edges).

pub fn pluck_edge(&mut self, edge: &EdgeRef) -> QueryGraphDependency

Removes the edge from the graph but leaves the graph intact by keeping the empty edge in the graph by plucking the content of the edge, but not the edge itself.

pub fn pluck_node(&mut self, node: &NodeRef) -> Node

Removes the node from the graph but leaves the graph intact by keeping the empty node in the graph by plucking the content of the node, but not the node itself.

pub fn remove_edge(&mut self, edge: EdgeRef) -> Option<QueryGraphDependency>

Completely removes the edge from the graph, returning it’s content. This operation is destructive on the underlying graph and invalidates references.

pub fn is_direct_child(&self, parent: &NodeRef, child: &NodeRef) -> bool

Checks if child is a direct child of parent.

Criteria for a direct child (either):

Every node that only has parent as their parent.

In case of multiple parents, all parents have already been visited before.

pub fn child_pairs(&self, node: &NodeRef) -> Vec<(EdgeRef, NodeRef), Global>

Returns a list of child nodes, together with their child edge for the given node. The list contains all children reachable by outgoing edges of node.

pub fn direct_child_pairs(
&self,
node: &NodeRef
) -> Vec<(EdgeRef, NodeRef), Global>

Returns all direct child pairs of node. See is_direct_child for exact definition of what a direct child encompasses.

pub fn zip_source_nodes(
&self,
edges: Vec<EdgeRef, Global>
) -> Vec<(EdgeRef, NodeRef), Global>

Resolves and adds all source NodeRefs to the respective EdgeRef.

pub fn mark_nodes(&mut self, parent_node: &NodeRef, child_node: &NodeRef)

Marks a node pair for swapping.