Function query_engine_tests::query_core::query_graph_builder::write::utils::emulate_on_delete_cascade

pub fn emulate_on_delete_cascade(
    graph: &mut QueryGraph, 
    relation_field: &Arc<RelationField>, 
    connector_ctx: &ConnectorContext, 
    _model: &Arc<Model>, 
    parent_node: &NodeRef, 
    child_node: &NodeRef
) -> Result<(), QueryGraphBuilderError>

Inserts cascade emulations into the graph between parent_node and child_node. relation_field is the relation field pointing to the model to be deleted. Recurses into the deletion emulation to ensure that subsequent deletions are handled correctly as well.

   ┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─
           Parent       │
   │  (ids to delete)    ─ ┐
    ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ┘
              │            │
              ▼
   ┌────────────────────┐  │
   │Find Connected Model│
┌──│     (Cascade)      │  │
│  └────────────────────┘
│             │            │
│             ▼
│┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ │
│  ┌────────────────────┐ │
││ │  Insert onDelete   │  │
│  │ emulation subtree  │ │
││ │ for all relations  │  │
│  │   pointing to D.   │ │
││ └────────────────────┘  │
│ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ┘
│             │            │
│             ▼
│  ┌────────────────────┐  │
└─▶│  Delete children   │
   └────────────────────┘  │
              │
              ▼            │
   ┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─
           Delete       │◀ ┘
   └ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─