Function query_engine_tests::query_core::query_graph_builder::write::utils::insert_existing_1to1_related_model_checks

pub fn insert_existing_1to1_related_model_checks(
    graph: &mut QueryGraph, 
    parent_node: &NodeRef, 
    parent_relation_field: &Arc<RelationField>
) -> Result<NodeRef, QueryGraphBuilderError>

Inserts checks and disconnects for existing models for a 1:1 relation. Expects that the parent node returns a valid ID for the model the parent_relation_field is located on.

Params: parent_node: Node that provides the parent id for the find query and where the checks are appended to in the graph. parent_relation_field: Field on the parent model to find children.

┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─
          Parent         │
└ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─
             :
             ▼
┌────────────────────────┐
│      Read related      │──┐
└────────────────────────┘  │
             │              │
             ▼              │
┌────────────────────────┐  │
│ If p > 0 && c. inlined │  │
└────────────────────────┘  │
        then │              │
             ▼              │
┌────────────────────────┐  │
│    Update ex. child    │◀─┘
└────────────────────────┘

The edge between Read Related and If fails on node count > 0 if child side is required.

We only need to actually update (“disconnect”) the existing model if the relation is also inlined on that models side, so we put that check into the if flow.

Returns a NodeRef to the “Read Related” node in the graph illustrated above.