Function query_engine_tests::query_core::query_graph_builder::write::utils::insert_find_children_by_parent_node

pub fn insert_find_children_by_parent_node<T>(
    graph: &mut QueryGraph, 
    parent_node: &NodeRef, 
    parent_relation_field: &Arc<RelationField>, 
    filter: T
) -> Result<NodeRef, QueryGraphBuilderError> where
    T: Into<Filter>, 

Adds a read query to the query graph that finds related records by parent ID. Connects the parent node and the read node with an edge, which takes care of the node transformation based on the parent ID.

Optionally, a filter can be passed that narrows down the child selection.

Returns a NodeRef to the newly created read node.

┌ ─ ─ ─ ─ ─ ─ ─ ─ ┐
      Parent
└ ─ ─ ─ ─ ─ ─ ─ ─ ┘
         │
         ▼
┌─────────────────┐
│  Read Children  │
└─────────────────┘

Example

Given two models, Blog and Post, where a blog has many posts, and a post has one block.

If the caller wants to query posts by blog ID:

parent_node needs to return a blog ID during execution.

parent_relation_field is the field on the Blog model, e.g. posts.

filter narrows down posts, e.g. posts where their titles start with a given string.