Crate query_engine_tests::query_core::schema

Modules

constants

Structs

DatabaseEnumType

Identifier

InputField

InputObjectType

ObjectType

OutputField

OutputType

QueryInfo

Designates a specific top-level operation on a corresponding model.

QuerySchema

The query schema defines which operations (query/mutations) are possible on a database, based on a Prisma schema.

Enums

EnumType

InnerOutputType

InputType

ObjectTag

Object tags help differentiating objects during parsing / raw input data processing, especially if complex object unions are present.

QueryTag

A QueryTag designates a top level query possible with Prisma.

ScalarType

Functions

build

build_with_features

compound_id_field_name

Computes a compound field name based on a multi-field id.

compound_index_field_name

Computes a compound field name based on an index.

itx_isolation_levels

Type Definitions

QuerySchemaRef