Module query_engine_tests::query_core::schema_builder

Query schema builder. Root for query schema building.

The schema builder creates all builders necessary for the process, and hands down references to the individual initializers as required.

Circular dependency schema building requires special consideration. Assume a data model looks like this, with arrows indicating some kind of relation between models:

      +---+
  +---+ B +<---+
  |   +---+    |
  v            |
+-+-+        +-+-+      +---+
| A +------->+ C +<-----+ D |
+---+        +---+      +---+

The above would cause infinite builder recursion circular dependency (A -> B -> C -> A) in relations (for example in filter building).

Without caching, processing D (in fact, visiting any type after the intial computation) would also trigger a complete recomputation of A, B, C.

Hence, all builders that produce input or output object types are required to implement CachedBuilder in some form to break recursive type building.

Additionally, the cache also acts as the component to prevent memory leaks from circular dependencies in the query schema later on, as described on the QuerySchema type. The cache can be consumed to produce a list of strong references to the individual input and output object types, which are then moved to the query schema to keep weak references alive (see TypeRefCache for additional infos).

Modules

constants

Enums

BuildMode

Build mode for schema generation.

Functions

append_opt

Appends an option of type T to a vector over T if the option is Some.

build

camel_case

Lowercases first letter, essentially. Assumes 1-byte characters, panics otherwise.

capitalize

Capitalizes first character. Assumes 1-byte characters.

compound_id_field_name

Computes a compound field name based on a multi-field id.

compound_index_field_name

Computes a compound field name based on an index.

field

Field convenience wrapper function.

init_input_object_type

Input object type initializer for cases where only the name is known, and fields are computed later.

input_field

Field convenience wrapper function.

input_object_type

Input object type convenience wrapper function.

object_type

Object type convenience wrapper function.

pluralize

Pluralizes given (English) input string. Falls back to appending “s”.

string_enum_type

Enum type convenience wrapper function.