Trait query_engine_tests::query_core::FilteredNestedMutation

pub trait FilteredNestedMutation {
    fn set_selectors(&mut self, selectors: Vec<SelectionResult, Global>);
}

Required Methods

fn set_selectors(&mut self, selectors: Vec<SelectionResult, Global>)

Sets the parent id of a nested mutation. This indicates the connector that it doesn’t need to refetch the child and that the mutation can directly be performed using that id.