Crate query_engine_tests::query_core

Modules

error

executor

What the executor module DOES:

interactive_transactions

interpreter

query_ast

query_document

Intermediate representation of the input document that is used by the query engine to build query ASTs and validate the incoming data.

query_graph

query_graph_builder

Query graph builder module. tbd

response_ir

Prisma Response IR (Intermediate Representation).

result_ast

schema

schema_builder

Query schema builder. Root for query schema building.

trace_helpers

Structs

AggregateRecordsQuery

Binding

CompactedDocument

ConnectRecords

CreateManyRecords

CreateRecord

DeleteManyRecords

DeleteRecord

DiffNode

DiffResult

Diff of two identifier vectors A and B: left contains all elements that are in A but not in B. right contains all elements that are in B but not in A.

DisconnectRecords

EdgeRef

Env

Expressionista

FieldConversionError

FieldCountError

FieldPair

ITXClient

ITXServer

In

IrSerializer

List

ManyRecordsQuery

NodeRef

OpenTx

ParsedArgument

ParsedField

ParsedInputMap

ParsedObject

QueryDocumentParser

QueryGraph

A graph representing an abstract view of queries and their execution dependencies.

QueryGraphBuilder

QueryInterpreter

QueryParserError

QueryPath

RawQuery

RecordAggregations

RecordQuery

RecordSelection

RelatedRecordsQuery

RelationViolation

ResponseData

Selection

SelectionBuilder

TransactionActorManager

TxId

How Interactive Transactions work The Interactive Transactions (iTx) follow an actor model design. Where each iTx is created in its own process. When a prisma client requests to start a new transaction, the Transaction Actor Manager spawns a new ITXServer. The ITXServer runs in its own process and waits for messages to arrive via its receive channel to process. The Transaction Actor Manager will also create an ITXClient and add it to hashmap managed by an RwLock. The ITXClient is the only way to communicate with the ITXServer. Once the prisma client receives the iTx Id it can perform database operations using that iTx id. When an operation request is received by the TransactionActorManager, it looks for the client in the hashmap and passes the operation to the client. The ITXClient sends a message to the ITXServer and waits for a response. The ITXServer will then perform the operation and return the result. The ITXServer will perform one operation at a time. All other operations will sit in the message queue waiting to be processed.

TxOpRequest

UpdateManyRecords

UpdateRecord

Enums

BatchDocument

CachedTx

Computation

ComputationResult

Conjuctive

CoreError

Expression

ExpressionResult

Flow

InterpreterError

Item

An IR item that either expands to a subtype or leaf-record.

Node

Operation

ParsedInputValue

Query

QueryDocument

QueryGraphBuilderError

QueryGraphDependency

Stored on the edges of the QueryGraph, a QueryGraphDependency contains information on how children are connected to their parents, expressing for example the need for additional information from the parent to be able to execute at runtime.

QueryGraphError

QueryParserErrorKind

QueryResult

QueryValue

ReadQuery

SelectionSet

TransactionError

TxOpRequestMsg

TxOpResponse

WriteQuery

Statics

CLOSED_TX_CACHE_SIZE

SHOW_ALL_TRACES

Traits

ArgumentListLookup

Builder

Temporary trait for the legacy read builder code.

FilteredQuery

QueryExecutor

TransactionManager

Functions

aggregate

db_name

execute_many_operations

execute_many_self_contained

execute_single_operation

execute_single_self_contained

extract_filter

Extracts a regular filter potentially matching many records.

extract_nested_rel_aggr_selections

extract_query_args

Expects the caller to know that it is structurally guaranteed that query arguments can be extracted, e.g. that the query schema guarantees that required fields are present. Errors occur if conversions fail.

extract_unique_filter

Extracts a filter for a unique selector, i.e. a filter that selects exactly one record.

find_first

find_many

find_related

find_unique

Builds a read query from a parsed incoming read query field.

format

format_expression

get_current_dispatcher

group_by

is_aggr_selection

is_user_facing_trace_filter

load

Loads a query executor based on the parsed Prisma schema (datasource).

set_parent_context_from_json_str

set_span_link_from_trace_id

spans_to_json

spawn_client_list_clear_actor

Spawn the client list clear actor It waits for messages from completed ITXServers and removes the ITXClient from the clients hashmap

spawn_itx_actor

Type Definitions

DataDependencyFn

ItemRef

Convenience type wrapper for Arc.

Map

A key -> value map to an IR item

ParsedInputList

ProjectedDataDependencyFn

QueryGraphBuilderResult

Query graph builder sub-result type.

QueryGraphResult

QueryParserResult

Result

Result type tying all sub-result type hierarchies of the core together.