Module query_engine_tests::query_core::executor

What the executor module DOES:

Defining an overarching executor trait, to be used on consumers of the core crate.

Defining executor implementations that combine the different core modules into a coherent string of actions to execute a given query document.

What the executor module DOES NOT DO:

Define low level execution of queries. This is considered an implementation detail of the modules used by the executors.

Traits

QueryExecutor

TransactionManager

Functions

db_name

execute_many_operations

execute_many_self_contained

execute_single_operation

execute_single_self_contained

get_current_dispatcher

load

Loads a query executor based on the parsed Prisma schema (datasource).