Module query_engine_tests::query_core::query_graph

Structs

DiffNode

EdgeRef

NodeRef

QueryGraph

A graph representing an abstract view of queries and their execution dependencies.

Enums

Computation

Flow

Node

QueryGraphDependency

Stored on the edges of the QueryGraph, a QueryGraphDependency contains information on how children are connected to their parents, expressing for example the need for additional information from the parent to be able to execute at runtime.

QueryGraphError

Functions

format

Type Definitions

DataDependencyFn

ProjectedDataDependencyFn

QueryGraphResult