Crate query_engine_tests::query_core::schema

Structs

ConnectorContext

Connector meta information, to be used in query execution if necessary.

DatabaseEnumType

Deprecation

FieldRefEnumType

Identifier

InputField

InputObjectType

InputObjectTypeConstraints

ObjectType

OutputField

QueryInfo

Designates a specific top-level operation on a corresponding model.

QuerySchema

The query schema. Defines which operations (query/mutations) are possible on a database, based on the (internal) data model.

StringEnumType

Enums

EnumType

InputType

ObjectTag

Object tags help differentiating objects during parsing / raw input data processing, especially if complex object unions are present.

OutputType

QueryTag

A QueryTag designates a top level query possible with Prisma.

ScalarType

Statics

MODEL_NAMESPACE

PRISMA_NAMESPACE

Traits

IntoArc

Since we have the invariant that the weak refs that are used throughout the query schema have to be always valid, we use this simple trait to keep the code clutter low.

QuerySchemaRenderer

Trait that should be implemented in order to be able to render a query schema.

Type Definitions

EnumTypeRef

EnumTypeWeakRef

InputFieldRef

InputObjectTypeStrongRef

InputObjectTypeWeakRef

ObjectTypeStrongRef

ObjectTypeWeakRef

OutputFieldRef

OutputTypeRef

QuerySchemaRef