Module query_engine_tests::schemas

Functions

a1_to_bm_opt

A <-1?—m-> B

all_composite_types

All types Prisma supports on a composite. Allows for simple picking of whatever should be tested from a single model.

autoinc_id

autoinc_id_cockroachdb

common_list_types

Test model containing all possible Prisma list types. Excludes capability-dependent types (e.g. JSON).

common_nullable_types

Test model containing all possible Prisma scalar types, nullable. Excludes capability-dependent types (e.g. JSON, Decimal).

common_numeric_types

Test model containing all common Prisma numeric types.

common_text_and_numeric_types

Test model containing all common Prisma numeric and string types.

common_text_and_numeric_types_optional

Test model containing all common Prisma numeric and string types, optional.

common_types

Test model containing all possible Prisma scalar types. Excludes capability-dependent types (e.g. JSON).

generic

Most basic datamodel containing only a model with ID for the most rudimentary testing.

json

Basic Test model containing a single json field.

json_default

Basic Test model containing a single optional json field.

json_opt

Basic Test model containing a single optional json field.

mixed_composites

Composites and relations mixed.

numeric_text_optional_one2m

posts_categories

User <-m—n-> posts

simple_m2m

string_combination

string_combination_unique

to_many_composites

Full to-many composite test model.

to_one_composites

Full to-one composite test model.

user

User model with some basic fields and unique constraints.

user_posts

User <-1—m-> posts