Module query_engine_tests::utils

Macros

assert_error

assert_query

is_one_of

jNull

Macro that replaces JSON nullability strings with the appropriate representation based on the connector capabilities. Allows us to reuse tests instead of copying.

match_connector_result

run_query

run_query_json

run_query_pretty

Enums

RawParam

Small abstraction over query raw parameters. Useful to differentiate floats from decimals as they’re encoded differently in clients.

Constants

TROUBLE_CHARS

Functions

date_iso_string

Convenience date string (UTC, RFC 3339) constructor so you don’t have to remember the format.

datetime_iso_string

Convenience full date + time string (UTC, RFC 3339) constructor so you don’t have to remember the format.

enclose

enclose_all

fmt_execute_raw

fmt_query_raw

now

Convenience date string (UTC, RFC 3339) constructor for the datetime right now, for cases you don’t care about the concrete DateTime in the tests.

string_to_base64

stringify